Break mock files into separate mock files
