+++ src/mock/mock-event.js
@@ -1,92 +1,6 @@
-import {getRandomArrayItem, getRandomPhoto, getRandomDescription} from '../utils.js';
-//import {POINT_TYPES} from '../const.js';
-
-const mockDestinations = [
-  {
-    id: 'dest1',
-    description: getRandomDescription(),
-    name: 'Chamonix',
-    pictures: [
-      {
-        src: getRandomPhoto(),
-        description: getRandomDescription(),
-      },
-      {
-        src: getRandomPhoto(),
-        description: getRandomDescription(),
-      },
-      {
-        src: getRandomPhoto(),
-        description: getRandomDescription(),
-      }
-    ],
-  },
-  {
-    id: 'dest2',
-    description: getRandomDescription(),
-    name: 'Amsterdam',
-    pictures: [],
-  },
-  {
-    id: 'dest3',
-    description: getRandomDescription(),
-    name: 'Jeneva',
-    pictures: [
-      {
-        src: getRandomPhoto(),
-        description: getRandomDescription(),
-      }
-    ],
-  }
-];
-
-const mockOffers = [
-  {
-    type: 'taxi',
-    offers: [
-      {
-        id: 't1',
-        title: 'Upgrade to a comfort class',
-        price: 120,
-      }
-    ]
-  },
-  {
-    type: 'flight',
-    offers: [
-      {
-        id: 'f1',
-        title: 'Business class food',
-        price: 40,
-      },
-      {
-        id: 'f2',
-        title: 'Extra luggage',
-        price: 80,
-      },
-      {
-        id: 'f3',
-        title: 'On-board TV-station',
-        price: 20,
-      }
-    ]
-  },
-  {
-    type: 'check-in',
-    offers: [
-      {
-        id: 'c1',
-        title: 'Mini-bar code',
-        price: 30,
-      },
-      {
-        id: 'c2',
-        title: 'Double bedroom',
-        price: 160,
-      },
-    ]
-  }
-];
+import {getRandomArrayItem} from '../utils.js';
+import {mockDestinations} from './mock-destination.js';
+import {mockOffers} from './mock-offers.js';
 
 const mockEvents = [
   {
