+++ src/model/events-model.js
@@ -1,5 +1,5 @@
 import {EVENT_COUNT} from '../const.js';
-import {getRandomEvent} from '../mock/trip-event.js';
+import {getRandomEvent} from '../mock/mock-event.js';
 
 export default class EventsModel {
   events = Array.from({length: EVENT_COUNT}, getRandomEvent);
